+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/util/RobotElementGenerator.java
@@ -11,6 +11,7 @@ import com.intellij.psi.impl.PsiFileFactoryImpl;
 import com.intellij.testFramework.LightVirtualFile;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotFeatureFileType;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotLanguage;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotConditionalContent;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotConditionalStructure;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotExceptionHandlingStructure;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotImportArgument;
@@ -21,7 +22,7 @@ import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotLoopControlStru
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotParameter;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotParameterId;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotPositionalArgument;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotPythonExpressionBody;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotPythonExpression;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotReturnStructure;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotTaskId;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotTemplateParameterId;
@@ -253,12 +254,12 @@ public record RobotElementGenerator(Project project) {
 
     public RobotLoopControlStructure createNewLoopControlStructure(LoopControlStructureType type) {
         String fileContent = """
-        *** Keywords ***
-        Dummy Keyword
-            FOR  ${item}  IN  @{items}
-                %s
-            END
-        """.formatted(type);
+                             *** Keywords ***
+                             Dummy Keyword
+                                 FOR  ${item}  IN  @{items}
+                                     %s
+                                 END
+                             """.formatted(type);
 
         PsiFile psiFile = createDummyPsiFile(fileContent);
         if (psiFile == null) {
@@ -375,7 +376,7 @@ public record RobotElementGenerator(Project project) {
         return returnStructureFinder.returnStructure;
     }
 
-    public RobotPythonExpressionBody createNewPythonExpressionBody(String pythonStatement) {
+    public RobotPythonExpression createNewPythonExpression(String pythonStatement) {
         String fileContent = """
                              *** Variables ***
                              ${Variable}=  ${{%s}}
@@ -385,9 +386,26 @@ public record RobotElementGenerator(Project project) {
         if (psiFile == null) {
             return null;
         }
-        RobotPythonExpressionBodyFinder pythonExpressionBodyFinder = new RobotPythonExpressionBodyFinder();
-        psiFile.acceptChildren(pythonExpressionBodyFinder);
-        return pythonExpressionBodyFinder.pythonExpressionBody;
+        RobotPythonExpressionFinder pythonExpressionFinder = new RobotPythonExpressionFinder();
+        psiFile.acceptChildren(pythonExpressionFinder);
+        return pythonExpressionFinder.pythonExpression;
+    }
+
+    public RobotConditionalContent createNewConditionalContent(String statement) {
+        String fileContent = """
+                             *** Keywords ***
+                             Dummy
+                                 IF  %s
+                                 END\\n
+                             """.formatted(statement);
+        PsiFile psiFile = createDummyPsiFile(fileContent);
+        if (psiFile == null) {
+            return null;
+        }
+
+        RobotConditionalContentFinder conditionalContentFinder = new RobotConditionalContentFinder();
+        psiFile.acceptChildren(conditionalContentFinder);
+        return conditionalContentFinder.conditionalContent;
     }
 
     public PsiFile createDummyPsiFile(String text) {
@@ -531,6 +549,16 @@ public record RobotElementGenerator(Project project) {
         }
     }
 
+    private static final class RobotConditionalContentFinder extends RecursiveRobotVisitor {
+
+        private RobotConditionalContent conditionalContent;
+
+        @Override
+        public void visitConditionalContent(@NotNull RobotConditionalContent o) {
+            conditionalContent = o;
+        }
+    }
+
     private static final class RobotExceptionHandlingStructureFinder extends RecursiveRobotVisitor {
 
         private RobotExceptionHandlingStructure exceptionHandlingStructure;
@@ -571,13 +599,13 @@ public record RobotElementGenerator(Project project) {
         }
     }
 
-    private static final class RobotPythonExpressionBodyFinder extends RecursiveRobotVisitor {
+    private static final class RobotPythonExpressionFinder extends RecursiveRobotVisitor {
 
-        private RobotPythonExpressionBody pythonExpressionBody;
+        private RobotPythonExpression pythonExpression;
 
         @Override
-        public void visitPythonExpressionBody(@NotNull RobotPythonExpressionBody o) {
-            pythonExpressionBody = o;
+        public void visitPythonExpression(@NotNull RobotPythonExpression o) {
+            pythonExpression = o;
         }
     }
 }